+++ ios/Flutter/flutter_export_environment.sh
@@ -1,12 +1,12 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=C:\flutter"
-export "FLUTTER_APPLICATION_PATH=F:\PROJECTS\context_menu"
+export "FLUTTER_ROOT=/home/salah/snap/flutter/common/flutter"
+export "FLUTTER_APPLICATION_PATH=/mnt/CEC0FE15C0FE038D/PROJECTS/flutter_context_menu"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
-export "FLUTTER_TARGET=lib\main.dart"
+export "FLUTTER_TARGET=lib/main.dart"
 export "FLUTTER_BUILD_DIR=build"
-export "FLUTTER_BUILD_NAME=0.0.1"
-export "FLUTTER_BUILD_NUMBER=0.0.1"
+export "FLUTTER_BUILD_NAME=0.1.1"
+export "FLUTTER_BUILD_NUMBER=0.1.1"
 export "DART_OBFUSCATION=false"
 export "TRACK_WIDGET_CREATION=true"
 export "TREE_SHAKE_ICONS=false"