+++ lib/flutter_context_menu.dart
@@ -1,7 +1,8 @@
-library flutter_context_menu;
-
-export 'src/entries/context_menu_entry.dart';
-export 'src/entries/context_menu_item.dart';
-export 'src/utils/helpers.dart';
+export 'src/components/default_context_menu_divider.dart';
+export 'src/components/default_context_menu_item.dart';
+export 'src/components/default_context_menu_text_header.dart';
+export 'src/core/models/context_menu_entry.dart';
+export 'src/core/models/context_menu_item.dart';
+export 'src/core/utils/helpers.dart';
 export 'src/widgets/context_menu.dart';
 export 'src/widgets/context_menu_state.dart';